requirements.txt
fastapi
uvicorn
requests
python-dotenv
gspread
google-auth